v1.011 Update dry mode

+++ services/local.py
@@ -15,19 +15,18 @@ def copy_local(path: str, args) -> None:
     """
     dest = LOCAL + '/' if not LOCAL.endswith('/') else LOCAL
     dest += get_file_name(path)
-    if args.dry:
+    file_exists = os.path.exists(dest)
+    if (not file_exists or args.override) and not args.dry:
+        logger.info(f'Выполняется копирование файла {path} в {dest}')
+        try:
+            shutil.copy2(path, dest)
+        except FileNotFoundError:
+            logger.error(f'Файл не найден {path}')
+        except IsADirectoryError:
+            logger.error(f'Каталог не найден {dest}')
+        else:
+            logger.info(f'Файл {path} успешно скопирован в {dest}')
+    elif file_exists and not args.override:
+        logger.info(f'Файл не скопирован, т.к. отключена перезапись')
+    elif args.dry:
         logger.info(f'Здесь могло быть копирование файла {path} в {dest}')
-    else:
-        file_exists = os.path.exists(dest)
-        if not file_exists or args.override:
-            logger.info(f'Выполняется копирование файла {path} в {dest}')
-            try:
-                shutil.copy2(path, dest)
-            except FileNotFoundError:
-                logger.error(f'Файл не найден {path}')
-            except IsADirectoryError:
-                logger.error(f'Каталог не найден {dest}')
-            else:
-                logger.info(f'Файл {path} успешно скопирован в {dest}')
-        elif file_exists and not args.override:
-            logger.info(f'Файл не скопирован, т.к. отключена перезапись')
